Add Basic .appveyor.yml With Linux Build
The AppVeyor system does not work in my fork of the project.  To make
the build common between forks and the main, a .yml file should be used.

Outline the sections and basics in some comments as described in the
AppVeyor getting started notes.  Copy the build steps from the current
.travis.yml file.

Once the Linux builds are working, I can start guessing at how to get
the Windows one running as well.