Check Inno Setup is installed
The build_installer.bat file uses Inno Setup to build the installer. It
does currently not check if Inno Setup is installed, and fails with the
message:
  The system cannot find the path specified.
  Build failed

That message does not explain how to fix the problem. I have added a
loop to search for Inno Setup 6 and 5. The user installed Inno Setup is
preferred over the system installed version. If Inno Setup cannot be
found then it prints an actionable error message.

I have tested it by:
1. Not having Inno Setup installed
2. Running build_installer.bat
3. Checked that the error message appears.
4. Installed Inno Setup 6 for my current user
5. Checked that the error message does not appear.
6. Checked that the installer package is generated
6. Uninstalled Inno Setup 6
7. Installed Inno Setup 6 for all users
8. Checked that the error message does not appear.
9. Checked that the installer package is generated
10. Uninstalled Inno Setup 6
11. Installed Inno Setup 5
12. Checked that the error message does not appear.
13. Checked that the installer package is generated