Ensure that window is not hidden with no system tray after startup
Before in all cases we would wait for the system tray for a period
of time, this would block users who don't have autohide enabled
and don't have a system tray from seeing the window.

Instead flip this around to either hide the window at startup if
autohide is enabled and show if it isn't. Then if we have opened
hidden wait a period of time and check if system tray is available,
if it isn't then force the window to be shown.

This provides much better UX for the general cases.

Closes #1024