Update Windows CI builds to multi-agent matrix
This PR updates Windows CI builds to use a multi-agent matrix when
building. Windows 2022 & 2019 have been DISABLED *temporarily* until
we can find a way to automagically detect the location of Visual
Studio build tools.

The only build agent available is Windows 2016, as it works with the
current setup. There is a publishing step, which only runs on the
Windows 2016 agent for now. This solves the issue of multiple
installer artifacts being built when the other agents are
re-enabled.

Per-commit generated binaries remain generated by all (not right now)
Windows build agenst.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>