Rename Barrier to Input Leap in translations
This is a bit hacky, but I ran `sed` on the translations, to change
'Barrier' to 'Input Leap', etc.

It seems to compile fine, but I haven't yet tested translations, as I'm
only fluent in English, unfortunately.

There are still references to `barriers` and `barrierc`, but we haven't
renamed those yet, so this will need more work.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>