dist/rpm: fix the specfile to work from a source package
A bunch of fixes required for this specfile to work at all:
- Spec files don't like the "-release" suffix in the Version field, so
  swap it to the numeric number only.
- Name and Urls needed updating
- Update some variables that still pointed to Synergy
- Sync the BuildRequires from the current Fedora barrier package
- Update the Source to use the default cmake tarball name
- Use the cmake rpm macros
- Add the new appdata file

Since clearly no-one is using this spec file as-is, we might as well
change it to work nicely on a local test. We can now run:

    mkdir build
    cd build
    cmake ..
    make package_source
    rpmbuild -D="%_sourcedir $PWD" -bb rpm/*.spec

To build an example RPM and check everything is still as expected.

The %{cmake} macros *may* be Fedora-specific, not sure. But since
apparently no-one has used this spec file for a while anyway it's
unlikely to be a huge issue.

Using cmake_install in particular means we no longer need the custom
install commands, we can just rely on the installed base.