dist/rpm: use the date/time/git sha as the rpm revision number
Instead of the mostly INPUTLEAP_BUILD_NUMBER which needs to be
externally supplied, use the date, time and git sha.

git revisions do not follow any sorting rules, so there's no guarantee
that two sequential git commits sort in the order of the commits. So
let's use a date and time for our test RPM packages, this way whichever
package was built last is going to be the most recent one as seen by
dnf/rpm.

The full NVR of the RPM Package built on July 1 at 06:30 UTC thus
becomes input-leap-2.4.0-202207010630.deadbeef.fc36.rpm. We can even
have this as an artefact of our CI infrastructure to allow for merge
request testing.