lib: notify the waiting thread when we delete the SocketMultiplexer
There's a race condition in the destruction of the SocketMultiplexer:
our thread is usually waiting on jobs_are_ready. If we destroy the
SocketMultiplexer, we cancel the thread and wait for it to finish, but
this never happens because the thread func never sees jobs_are_ready
change.

Fix this by pretending jobs are ready.

Fixes #1490