Drop support for --no-xinitthreads
This was added 11 years ago but without information in the commit as to
why it was added. Likely there was a bug in some Xlib version that
required this.

This bug was probably fixed a long time ago and as of libX11 1.8 libX11
now always calls XInitThreads anyway. So let's drop this code.