Update gtest to 1.12.0, drop standalone gmock
Bump gtest to the latest version, primarily so we can get the
GTEST_SKIP() macro that allows us to skip tests dynamically.

As of version 1.8.0, the standalone gmock is deprecated, that is now
part of gtest, so gest sources are in ext/gtest/googletest and gmock is
in ext/gtest/googlemock.