test: if we failed to open the display and we don't have DISPLAY, skip
Running in the CI we may not have an X server running, so let's not fail
on an failed XOpenDisplay, just skip it. Only fail if we *do* have
DISPLAY set and XOpenDisplay failed anyway.

At this point, no standard distribution setup should ever have DISPLAY
at NULL anyway and for local custom setups we should encourage the user
to set DISPLAY.

This affects two failing cases, one in the XWindowsScreenTests (where
failing to open the display exits the XWindowsScreen constructor early
and thus fails the EXPECT_CALL) and the XWindowsKeyStateTests where
failing to open the display asserted.