treewide: drop the whitespace alignment for class methods
Realign all header files to use a single space between the method return
type and its name, i.e. "void foo();" instead of the previous attempted
alignment of all methods to the same tabstop.

Patch created by running:
    sed -i 's/[ ]\+\([a-z]\w*\)(/ \1(/' src/**/*.h
and then fixing/ignoring the outliers.

This does not align the method arguments, but it's an improvement
nonetheless.