lib: make the parent_requests_shutdown virtual
And override it with an always-false in the TestEventQueue.

This fixes the test case failure in the IpcTests when run in the github
CI. Our NonBlockingStream() is initialized with stdin, but this gives us
EINPROGRESS followed by ENOENT when trying to read from it - presumably
due to some setup within the github CI.

Fix this by assuming that we do not have a parent process in the tests,
so it cannot possibly requests us to shut down.