dist: Add basic support for flatpaking InputLeap
Provide a shell wrapper "input-leap-flatpak" that checks for one of two
configuration files. If the $XDG_CONFIG_HOME/InputLeap/server.name file
exists, it is used as server location for input-leapc. (Note that
flatpak doesn't allow mDNS so you can't use foo.local)

Otherwise, if InputLeap.conf exists in that directory we start the
server with that config file. Note that input-leaps is untested because
it requires the xdg-desktop-portal and various other bits in place which
I don't have all ready yet.

The script overrides the LIBEI_SOCKET environment variable to be in
XDG_RUNTIME_DIR/eis-0, so placing a socket there will enable the
existing hacks to connect directly.

Modified-by: Povilas Kanapickas <povilas@radix.lt>