ei: Fix EiKeyState::getKeyMap() key/modifiers combo
InputLeap does its best to replicate a key + modifier to achieve the
desired symbol.

To do, InputLeap considers the modifiers required to get the symbol, but
also the modifiers the keyval is sensitive to.

getKeymap() would however add only one key item entry for all the
required modifiers, which later breaks the logic in the keymap code
KeyMap::keysForKeyItem() as it would try to press all the modifiers
to get the desired symbol.

In the case of Shift, it also meant pressing CapsLock, which obviously
breaks the logic, as pressing Shift while CapsLock is enabled would
produce the lowercase symbol.

Fix EiKeyState::getKeyMap() to add a different key entry for all the
required modifiers.

Modified-by: Povilas Kanapickas <povilas@radix.lt>