ei: Release input capture when motion is on primary
Our implementation of input capture sets barriers on the edges of all
the zones, regardless of the actual machines layout.

That means that input capture will activate every time the pointer
reaches the edge of the screen. When there is no secondary screen found
at the given edge, we might either jump to the wrong secondary screen
or worse, stop pointer input entirely.

To avoid this, simply check whether the input capture is active while
we're still on the primary screen, and disable input capture when that's
the case.

Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>
Modified-by: Povilas Kanapickas <povilas@radix.lt>