lib/clipboard: Do not try to use clipboard if it failed
Server::onClipboardChanged() tries to get the clipboard from the
sender by calling getClipboard().

getClipboard() however may fail, and return false in that case, yet
onClipboardChanged() does not actually check the return value and tries
to use the clipboard.

If getClipboard() does nothing as with the EiScreen implementation,
bad things can happen, such as an abort of Input Leap on an assertion
failure in the marshall() call that follows.

To avoid that annoyance, check the return value from getClipboard() and
bail out early (and gracefully) when it's reported as false.

Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>