ei: Reset XKB state on keymap change
We keep the XKB state around, but the XKB state depends on the keymap.

At startup, we create a default keymap, then later replaced by the
actual keymap.

However, we would not update the XKB state when doing so, and we would
still be using the old (default US), keymap.

To solve the problem, make sure to recreate both the keymap and the XKB
state in sync.

Modified-by: Povilas Kanapickas <povilas@radix.lt>