ei: Pass the keymap as string, not as buffer
libxkbcommon fails to compile a keymap that contains a terminating null
byte if passed as buffer. Ensure we always have a null-terminated string
instead and pass it as such to the keymap compiler. This avoids issues
when different EIS implementation give us the keymap with or without
that trailing null.

Modified-by: Povilas Kanapickas <povilas@radix.lt>