lib: Don't shutdown app when debugger interrupts it
The accepted convention to interrupt program that gdb is attached to is
by calling Ctrl+C in gdb console. This sends SIGINT which is then
intercepted by gdb. However, signals in InputLeap apps are intercepted
by sigwait(). This does not allow gdb to intercept them first. As a
result, interrupting app via Ctrl+C causes the program to terminate as
it thinks it received SIGINT.

This is worked around by checking if app is under debugger and raising
SIGTRAP instead if that's the case.