Fix conflicting artefact names
This is the first commit in this PR that fixes the `upload-artifact` error
from Azure (HTTP 503), causing the check to fail.

We now upload the ZIP artefact in the format of `${windows-ver}-Release`.
This makes the artefact name *unique*, and solves the first problem.

The fix is inspired by: https://github.com/actions/upload-artifact/issues/171#issuecomment-1007504385