lib: Fix potential ODR rule violation due to EventQueueTimer
Currently every platform defines its own EventQueueTimer type. This
assumes that only one platform can be enabled at a time. In the future
this will not necessarily be the case (Ei platform may be enabled
together with XWindows). Depending on exact implementation choices this
may cause ODR violations.

It's better to expose such an API that extending it in the expected way
does not have chance for ODR violations. Having just one EventQueueTimer
type is the simpliest solution.