base: Fix race condition in EventQueue event dispatching
Currently EventQueue takes a lock on event handlers tables when
modifying or accessing them. The access in get_handler() returns a
pointer to a handler which is no longer protected by a lock. It is
possible that the handler is removed from the event handler tables
before it is invoked in dispatchEvent() resulting in access after free
bug.