clipboard: block updates when sender is not the owner
This blocks clipboard updates when the sender is not the clipboard
owner.

This is actually being done in an assertion in onClipboardChanged, but
obviously these are removed from production builds. As such, this is
effectively allowing clipboard updates from any screen, no matter if
it's the current screen or not, which breaks the clipboard state.  This
can cause crashes of the server if the screen is moved at certain times
during this process.

Note that this has a side-effect of possibly blocking clipboard updates
if the screen is switched before a clipboard update is acknowledged.
This delay is obviously quite short, so observing it is hopefully rare.

Fixes #1593.