fix: wake up on mac when screen locked
@shymega's note: from debauchee/barrier#1927, thank you @junbujianwpl.

I'm creating PRs from Barrier's unmerged PRs for Input Leap. This is so
we can a) let the community know that there is an active fork and b)
take advantage of the code contributions that Barrier simply isn't
accepting. This gives us the benefit of improved codebase and bug fixes,
such as this one, which happens on Mac machines.