Revert "only allows active client grabbing clipboard"
This reverts commit 9ad9b3a322e9559e947633e15989cebafef3accb.

Based on discussion in #1624 it seems like this would be better denoted
as intended behavior and if a truly shared clipboard was not desired,
that clipboard sharing should be disabled.

Blocking off-screen updates truly is a complex process and there are
other parts of the stack that need to be fixed as well (see
aforementioned issue). If this is a feature that is desired, a larger
update will be needed, and likely should also be introduced as an
optional feature.