Allow Wayland warning to be switched off at compile-time
This commit creates a CMake 'option' for disabling/enabling the Wayland
warning. It is enabled only under Unix, non-Apple targets, and by
default, the warning is enabled. It also depends on the
`INPUTLEAP_BUILD_LIBEI` option, as it's nested under the logic for that
option.

The reason for this commit was to allow users to override the otherwise
fine, automatic conditionals detecting Wayland support.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>