fix(gui/$DISPLAY): Check for X11 on the Input Leap GUI with Qt
As reported in #1792, a bug has been found where Input Leap GUI doesn't
recognise 'libportal' support as being present on the system, and thus
refuses to start.

The idea behind this PR only works on the GUI, so we need a long-term
solution, but currently, it replaces the X11 check with
a call to `QGuiApplication::platformName` instead.

We test for either `xcb` or `wayland`. If X11 is running, we do nothing.

If Wayland is running, I moved the existing Wayland check, to the new
`if` block in this commit, and it now displays an Information msgbox to
inform the user about current Wayland support.

Fixes #1792.
Closes #1794.

This PR and #1794 are mutually exclusive, and only one is needed at this
point. Both address #1792, but I've approached the problem from two
different perspectives.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>