fix(gui/$DISPLAY): Add guard to X11/XWayland `$DISPLAY` check
As reported in #1792, a bug has been noticed where Input Leap GUI
doesn't recognise 'libportal' support as being present on the system,
and thus refuses to start.

This is a compile-time guard, so it's not checked at run-time.

In future, I will refactor this check to be optional (enabled by
default, adjustable by CMake option), and to simply probe libportal
and/or X11 safely, to ascertain whenever it needs the `$DISPLAY`
variable. It is not a perfect fix, and we should look to refactor.

Fixes #1792.
Closes #1800 - this PR and #1800 are mutually exclusive, and only one is
needed at this point. Both address #1792, but I've approached the
problem from two different perspectives.