ei: fix scroll events
We may get scroll events or scroll discrete events and both may be in
fractions of a single logical scroll event. Handle this by mapping 10
pixels to one wheel click (that's what mutter/gtk do too) and otherwise
keeping the remainder around in the device for the next scroll event.

Note that we're always sending things through as wheel, so on
the client side we only ever call ei_device_scroll_discrete().

In theory, our pointer and keyboard device may be the same ei device, so
let's assume we have a user data on all of them and free them during
cleanup - that way we don't need complicated checks of which devices are
the same device.

Fixes #1840