Change from LOG(CLOG_WARN ...) to LOG_WARN(...) and friends
Drop the hack that used the first characters of the format string
to stash the priority and instead provide several different macros that
expand to a proper function call passing the priority through.

This is a bit harder to do with LOGC but there are only three calls of
that in the code so let's drop that macro altogether.