ei: implement server-side hotkey handling
Add a map that contains the key id and a vector of modifier masks + the
assigned hotkey id, then register and lookup those hotkeys as we get the
events. This is a bit minimum effort but should get us 80% of the way
there, further inspiration can be taken from the X implementation of the
same thing.

Note that due to how the InputCapture portal works, this code is not
triggered unless we're currently capturing input - and that only happens
when we're logically on the remote.

No solution currently exists in the ei backend to capture the key on the
server side.

Fixes #1711 (partially)