ei: start with an initial screen size of 1x1
getShape() is called as soon as we connect to the server but at this
point we do not have a shape yet, we're still waiting for the user to
confirm the portal request. If we return all zeroes here the server
disconnects us (ClientProxy1_6::recvInfo()), resulting in a continuous
connect/disconnect loop until the user finally says yes.

Send an intial screen shape instead, that shape will get updated later
when we get the devices through libei.

Fixes #1718