ei: Send the updated shape to the server
In commit c9978941 we now initialize with a screen shape of 1x1 and then
update the screen shape later when we get the regions through libei. But
we never sent that updated screen shape to the server so it kept
thinking we're a teensy-weensy one-pixel screen.

Fixes c9978941de55de8ed2b6a4872c427c7765ad85da