github: Remove opensuse/tumbleweed from the build set
Tumbleweed has problems installing dependencies lately (most likely due
to glibc upgrade and rebuild). The Docker repository does not provide
any specific tags, only latest. As a result, Tumbleweed cannot be
depended on for CI builds.

This commit should be reverted as soon as OpenSuse Leap 16 is released
or tumbleweed is fixed.