Update README.md
Added a note about the AltGr problem that will affect anybody with eu keyboards that uses a linux server and windows client. I even started looking at the source code before realizing there's a workaround, maybe adding the workaround on the readme is also a good idea so people doesn't have to read 100 lines of issue to get to it.